created class for testing db functionality
