+++ surveylyzer-backend/src/main/java/ch/zhaw/surveylyzerbackend/SurveylyzerBackendApplication.java
@@ -1,5 +1,6 @@
 package ch.zhaw.surveylyzerbackend;
 
+import ch.zhaw.db.TmpTestDB;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
@@ -7,16 +8,19 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 
 @SpringBootApplication
-@ComponentScan(basePackages = {"ch.zhaw.pdfReceiver"})
+//@ComponentScan(basePackages = {"ch.zhaw.pdfReceiver"})
+@ComponentScan(basePackages = {"ch.zhaw"})
 @EnableMongoRepositories("ch.zhaw.domain")
 public class SurveylyzerBackendApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(SurveylyzerBackendApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(SurveylyzerBackendApplication.class, args);
+    }
 
-	@Bean
-	public TmpTestDB helloRunner() {
-		return new TmpTestDB();
-	}
+
+    // Todo: remove running db-test if not needed anymore...
+    @Bean
+    public TmpTestDB helloRunner() {
+        return new TmpTestDB();
+    }
 }
